Split admin UI into components
