+++ client/web/src/enterprise/insights/admin-ui/components/job-actions/index.ts
@@ -1,0 +1,1 @@
+export { CodeInsightsJobsActions } from './CodeInsightsJobsActions'
