+++ client/web/src/enterprise/insights/admin-ui/components/job-card/CodeInsightsJobCard.module.scss
@@ -1,0 +1,117 @@
+.insight-job {
+    display: grid;
+    grid-template-columns: min-content auto min-content;
+    grid-template-areas: 'checkbox content state';
+    align-items: baseline;
+    padding: 0.75rem 0.75rem 0.75rem 0.5rem;
+    border-bottom: 1px solid var(--border-color);
+
+    &--active {
+        border-left: 3px solid var(--primary);
+        background-color: var(--gray-01);
+
+        :global(.theme-dark) & {
+            background-color: var(--gray-09);
+        }
+    }
+
+    &-checkbox {
+        grid-area: checkbox;
+        margin: 0 1rem 0 0.25rem;
+    }
+
+    &-content {
+        grid-area: content;
+        display: flex;
+        flex-direction: column;
+        align-items: flex-start;
+    }
+
+    &-header {
+        display: flex;
+        flex-wrap: wrap;
+        align-items: baseline;
+        gap: 0.25rem;
+    }
+
+    &-title {
+        margin-bottom: 0.25rem;
+        margin-right: 0.5rem;
+        font-size: 1.25rem;
+        cursor: pointer;
+    }
+
+    &-subtitle {
+        margin-bottom: 0.25rem;
+    }
+
+    &-main-info {
+        margin-top: 0.25rem;
+        display: flex;
+        flex-direction: column;
+    }
+
+    &-query-block {
+        line-height: 1.8;
+    }
+
+    &-query {
+        padding: 0.15rem;
+        margin-left: 0.5rem;
+        color: var(--body-color);
+        background: var(--gray-02);
+
+        :global(.theme-dark) & {
+            background-color: var(--color-bg-3);
+        }
+    }
+
+    &-state {
+        grid-area: state;
+        display: flex;
+        gap: 0.25rem;
+        align-self: flex-start;
+        color: var(--text-muted);
+
+        &--completed {
+            color: var(--success);
+        }
+
+        &--errored {
+            color: var(--danger);
+        }
+
+        &--queued {
+            color: var(--primary);
+        }
+    }
+}
+
+.pill {
+    font-weight: normal;
+
+    :global(.theme-dark) & {
+        background-color: var(--color-bg-3);
+    }
+
+    :global(.theme-light) & {
+        background-color: var(--oc-gray-1);
+    }
+}
+
+.errors-trigger {
+    padding: 0.2rem 0.5rem;
+}
+
+.errors-content {
+    padding: 0.5rem;
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+
+.error {
+    margin: 0;
+    flex-shrink: 0;
+    max-width: 40rem;
+}
