+++ client/web/src/enterprise/insights/admin-ui/components/job-card/index.ts
@@ -1,0 +1,1 @@
+export { CodeInsightsJobCard } from './CodeInsightsJobCard'
